+++ navground_sim_py/navground/sim/tasks/__init__.py
@@ -1,1 +1,2 @@
 from .._navground_sim import DirectionTask, WaypointsTask
+from .path import PathTask